requirements.txt
flask
opencv-python
tensorflow
opencv-python
torch
pillow
pandas
fal_client
requests
flask_pymongo
flask_session